requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
